refactor(version): move version logic to pkg
