+++ .gitignore
@@ -2,3 +2,6 @@
 
 dist/
 config/
+
+/ptparchiver
+/ptparchiver.exe
